feat(reviews): Task(5-7)

+++ backend/ReviewService/ReviewService.Application/Features/Reviews/Commands/DeleteReview/DeleteReviewCommandHandler.cs
@@ -1,12 +1,10 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
+using System.Threading;
 using System.Threading.Tasks;
 using MediatR;
 using ReviewService.Application.Interfaces.Commands;
 using ReviewService.Domain.Entities;
-using ReviewService.Domain.Helpers;
+using ReviewService.Domain.Helpers; // Для NotFoundException
 using ReviewService.Domain.Interfaces;
 
 namespace ReviewService.Application.Features.Reviews.Commands.DeleteReview
@@ -24,6 +22,16 @@ namespace ReviewService.Application.Features.Reviews.Commands.DeleteReview
         {
             Review? review = await reviewRepository.GetByIdAsync(request.ReviewId, cancellationToken);
             if (review == null) throw new NotFoundException($"Review with ID {request.ReviewId} not found!");
+
+            // --- ЗАВДАННЯ 7: Перевірка прав ---
+            bool isAdmin = request.RequesterRole == "Admin";
+            bool isOwner = review.User.UserId == request.RequesterId;
+
+            if (!isOwner && !isAdmin)
+            {
+                throw new UnauthorizedAccessException("You do not have permission to delete this review.");
+            }
+            // ----------------------------------
 
             await reviewRepository.DeleteAsync(request.ReviewId, cancellationToken);
             return Unit.Value;
